clean up the public api
